packages/squad/requirements.txt
--index-url http://david.ramirez:12345678@10.43.102.250:3000/api/packages/heimdall/pypi/simple
--extra-index-url https://pypi.org/simple
--trusted-host 10.43.102.250
modi >= 0.1.2
fastapi >= 0.111.0
uvicorn >= 0.30.1
python-dotenv >= 1.0.1
fastapi-router-controller >= 0.5.0
